removed tasks app completely and merged student table in user table
