+++ students/forms.py
@@ -1,22 +1,14 @@
 from django import forms
 from django import forms
 from django.contrib.auth import get_user_model
-
-from tasks.models import Student
-
+from django.contrib.auth.forms import UserCreationForm, UsernameField
 
 User = get_user_model()
 
 
-profile_choices = (
-    ("Developer", "Developer"),
-    ("AI", "AI")
-)
-
-
 class StudentModelForm(forms.ModelForm):
     class Meta:
-        model = Student
+        model = User
         fields = (
             'first_name',
             'last_name',
@@ -37,7 +29,6 @@ class StudentModelForm(forms.ModelForm):
             'joining_date': forms.DateInput(format=('%Y-%m-%d'), attrs={'class':'form-control', 'placeholder':'Select a date', 'type':'date'}),
             'offer_letter_issue_date': forms.DateInput(format=('%Y-%m-%d'), attrs={'class':'form-control', 'placeholder':'Select a date', 'type':'date'}),
             'completion_letter_issue_date': forms.DateInput(format=('%Y-%m-%d'), attrs={'class':'form-control', 'placeholder':'Select a date', 'type':'date'}),
-            # 'Profile': forms.MultipleChoiceField(choices=profile_choices)
         }
 
 
@@ -47,10 +38,17 @@ class NewStudentModelForm(forms.ModelForm):
         fields = (
             'first_name',
             'last_name',
-            'username',
+            'college_roll_number',
             'phone_number',
             'email'
         )
         labels = {
-            "username": "roll number"
+            "college_roll_number": "roll number"
         }
+
+
+class CustomUserCreationForm(UserCreationForm):
+    class Meta:
+        model = User
+        fields = ("username",)
+        field_classes = {'username': UsernameField}
