+++ students/templates/students/student_update.html
@@ -7,7 +7,7 @@
         <div class="lg:w-3/5 mx-auto flex flex-wrap">
         <div class="w-full lg:pr-10 lg:py-6 mb-6 lg:mb-0">
             <h2 class="text-sm title-font text-gray-500 tracking-widest">STUDENT</h2>
-            <h1 class="text-gray-900 text-3xl title-font font-medium mb-4">{{ task.task_name }}</h1>
+            <h1 class="text-gray-900 text-3xl title-font font-medium mb-4">{{ student }}</h1>
             <div class="flex mb-4">
             <a href="{% url 'students:student-detail' student.pk %}" class="flex-grow border-b-2 border-gray-300 py-2 text-lg px-1">Details</a>
             <a class="flex-grow text-indigo-500 border-b-2 border-indigo-500 py-2 text-lg px-1">Update</a>
