+++ students/views.py
@@ -2,17 +2,26 @@ from django.contrib.auth import get_user_model
 from django.shortcuts import render, reverse
 from django.views import generic
 from .mixins import AdminAndLoginRequiredMixin
-
-from tasks.models import Student
-from .forms import NewStudentModelForm, StudentModelForm
+from .forms import NewStudentModelForm, StudentModelForm, CustomUserCreationForm
 
 User = get_user_model()
+
+
+class SignupView(generic.CreateView):
+    template_name = "registration/signup.html"
+    form_class = CustomUserCreationForm
+
+    def get_success_url(self):
+        return reverse("login")
+
 
 class StudentListView(AdminAndLoginRequiredMixin, generic.ListView):
     template_name = "students/student_list.html"
 
     def get_queryset(self):
-        return Student.objects.all()
+        queryset = User.objects.all()
+        queryset = queryset.filter(is_admin = False)
+        return queryset
 
 
 class StudentCreateView(AdminAndLoginRequiredMixin, generic.CreateView):
@@ -24,18 +33,10 @@ class StudentCreateView(AdminAndLoginRequiredMixin, generic.CreateView):
 
     def form_valid(self, form):
         user = form.save(commit=False)
+        user.username = f"{user.first_name}_{user.college_roll_number}"
         user.is_admin = False
         user.set_password(user.phone_number)
         user.save()
-
-        Student.objects.create(
-            user=user,
-            first_name=user.first_name,
-            last_name=user.last_name,
-            phone_number=user.phone_number,
-            email=user.email,
-            college_roll_number=user.username
-        )
 
         return super(StudentCreateView, self).form_valid(form)
 
@@ -45,15 +46,20 @@ class StudentDetailView(AdminAndLoginRequiredMixin, generic.DetailView):
     context_object_name = "student"
 
     def get_queryset(self):
-        return Student.objects.all()
+        queryset = User.objects.all()
+        queryset = queryset.filter(is_admin = False)
+        return queryset
 
 
 class StudentUpdateView(AdminAndLoginRequiredMixin, generic.UpdateView):
     template_name = "students/student_update.html"
     form_class = StudentModelForm
+    context_object_name = "student"
 
     def get_queryset(self):
-        return Student.objects.all()
+        queryset = User.objects.all()
+        queryset = queryset.filter(is_admin = False)
+        return queryset
 
     def get_success_url(self):
         return reverse("students:student-list")
@@ -64,7 +70,9 @@ class StudentDeleteView(AdminAndLoginRequiredMixin, generic.DeleteView):
     context_object_name = "student"
 
     def get_queryset(self):
-        return Student.objects.all()
+        queryset = User.objects.all()
+        queryset = queryset.filter(is_admin = False)
+        return queryset
 
     def get_success_url(self):
         return reverse("students:student-list")
