--- tasks/templates/tasks/task_create.html
@@ -1,18 +1,0 @@
-{% extends 'base.html' %}
-{% load tailwind_filters %}
-
-{% block content %}
-    <div class="max-w-lg mx-auto">
-        <a class="hover:text-blue-500" href="{% url 'tasks:task-list' %}">Go back to tasks</a>
-        <div class="py-5 border-t border-gray-200">
-            <h1 class="text-4xl text-gray-700">Create a new task</h1>
-        </div>
-        <form method="post" class="mt-5">
-            {% csrf_token %}
-            {{ form|crispy }}
-            <button type='submit' class="w-full text-white bg-blue-500 hover:bg-blue-600 px-3 py-2 rounded-md">
-                Submit
-            </button>
-        </form>
-    </div>
-{% endblock content %}
