--- tasks/templates/tasks/task_detail.html
@@ -1,36 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<section class="text-gray-600 body-font overflow-hidden">
-  <div class="container px-5 py-24 mx-auto">
-    <div class="lg:w-3/5 mx-auto flex flex-wrap">
-      <div class="w-full lg:pr-10 lg:py-6 mb-6 lg:mb-0">
-        <h2 class="text-sm title-font text-gray-500 tracking-widest">TASK</h2>
-        <h1 class="text-gray-900 text-3xl title-font font-medium mb-4">{{ task.task_name }}</h1>
-        <div class="flex mb-4">
-          <a class="flex-grow text-indigo-500 border-b-2 border-indigo-500 py-2 text-lg px-1">Details</a>
-          <a href="{% url 'tasks:task-update' task.pk %}" class="flex-grow border-b-2 border-gray-300 py-2 text-lg px-1">Update</a>
-        </div>
-        <div class="flex border-t border-gray-200 py-2">
-          <span class="text-gray-500 pr-5">Project name</span>
-          <span class="ml-auto text-gray-900">{{ task.project_name }}</span>
-        </div>
-        <div class="flex border-t border-gray-200 py-2">
-          <span class="text-gray-500 pr-5">Task name</span>
-          <span class="ml-auto text-gray-900">{{ task.task_name }}</span>
-        </div>
-        <div class="flex border-t border-b mb-6 border-gray-200 py-2">
-          <span class="text-gray-500 pr-5">Description</span>
-          <span class="ml-auto text-gray-900">{{ task.description }}</span>
-        </div>
-        <div class="flex">
-            <a href="{% url 'tasks:task-delete' task.pk %}" class="flex mx-auto text-white bg-indigo-500 border-0 py-2 px-6 focus:outline-none hover:bg-indigo-600 rounded">
-                Delete Task
-            </a>
-        </div>
-      </div>
-    </div>
-  </div>
-</section>
-{% endblock content %}
-
