--- tasks/templates/tasks/task_list.html
@@ -1,44 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<section class="text-gray-600 body-font">
-  <div class="container px-5 py-24 mx-auto flex flex-wrap">
-    <div class="w-full mb-6 py-6 flex justify-between items-center border-b border-gray-200">
-      <div>
-        <h1 class="text-4xl text-gray-800">Tasks</h1>
-      </div>
-      {% if not request.user.is_admin %}
-        <div>
-          <a class="text-gray-500 hover:text-blue-500" href="{% url 'tasks:task-create' %}">create task</a>
-        </div>
-      {% endif %}
-    </div>
-    <div class="flex flex-wrap -m-4">
-      {% for task in tasks %}
-        <div class="p-4 lg:w-full md:w-full">
-          <div class="flex border-2 rounded-lg border-gray-200 border-opacity-50 p-8 sm:flex-row flex-col">
-            <div class="w-16 h-16 sm:mr-8 sm:mb-0 mb-4 inline-flex items-center justify-center rounded-full bg-indigo-100 text-indigo-500 flex-shrink-0">
-              <svg fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" class="w-8 h-8" viewBox="0 0 24 24">
-                <path d="M22 12h-4l-3 9L9 3l-3 9H2"></path>
-              </svg>
-            </div>
-            <div class="flex-grow">
-              <h2 class="text-gray-900 text-lg title-font font-medium mb-3">Project : {{ task.project_name }}</h2>
-              <p class="leading-relaxed text-base">
-                Task name : {{ task.task_name }}
-              </p>
-              <a href="{% url 'tasks:task-detail' task.pk %}" class="mt-3 text-indigo-500 inline-flex items-center">
-                View Task
-                  <svg fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" class="w-4 h-4 ml-2" viewBox="0 0 24 24">
-                    <path d="M5 12h14M12 5l7 7-7 7"></path>
-                  </svg>
-              </a>
-            </div>
-          </div>
-        </div>
-      {% endfor %}
-    </div>
-  </div>
-</section>
-{% endblock content %}
-
