+++ templates/navbar.html
@@ -26,7 +26,7 @@
           <a href="#" class="mr-5 hover:text-gray-900">Offer Letter</a>
           <a href="#" class="mr-5 hover:text-gray-900">Completion Letter</a>
         {% endif %}
-        <a href="{% url 'tasks:task-list' %}" class="mr-5 hover:text-gray-900">Tasks</a>
+        {% comment %} <a href="{% url 'tasks:task-list' %}" class="mr-5 hover:text-gray-900">Tasks</a> {% endcomment %}
       {% endif %}
     </nav>
     {% if request.user.is_authenticated %}
